backend/requirements.txt
Django==5.2.8
djangorestframework==3.15.0
django-cors-headers==4.0.0
channels==4.0.0
psycopg2-binary>=2.9
gunicorn>=20.1.0
whitenoise>=6.5.0
# Dev / tooling
black==24.3.0
ruff==0.6.1
pytest-django>=4.5.0
